requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-session==0.5.0
mysql-connector-python==8.1.0
bcrypt==4.0.1
python-dotenv==1.0.0
pytest==7.4.2
pytest-flask==1.2.0
